# PERF: extend SSE2 operations to anynan/allnan in contiguous axis case