Some slight slowdowns in the other axis for `gcc7`, but otherwise broad speedups:
```
$ asv compare HEAD^ HEAD -s --sort ratio  --only-changed | grep -v fast | grep -v int
       before           after         ratio
     [3d1cb0a2]       [0386185c]
-         652±8μs          515±9μs     0.79  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        651±30μs         506±10μs     0.78  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        665±20μs         508±20μs     0.76  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        772±10μs        578±100μs     0.75  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        753±10μs         555±40μs     0.74  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        752±10μs         536±20μs     0.71  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        783±10μs         547±40μs     0.70  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        743±10μs         510±10μs     0.69  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-         769±6μs         517±10μs     0.67  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        762±20μs          507±9μs     0.67  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        762±20μs          504±8μs     0.66  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        777±30μs          503±6μs     0.65  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        814±60μs         525±20μs     0.65  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        798±30μs          499±8μs     0.63  reduce.TimeAllNan2D.time_allnan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-       843±200μs          524±7μs     0.62  reduce.TimeAnyNan2D.time_anynan('float64', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-       560±200μs         342±90μs     0.61  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        521±10μs          290±8μs     0.56  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        532±20μs          288±7μs     0.54  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-         512±5μs          269±8μs     0.52  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-         511±7μs          259±3μs     0.51  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        520±10μs          260±4μs     0.50  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        661±40μs         301±70μs     0.46  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        665±20μs         291±80μs     0.44  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        662±20μs          260±5μs     0.39  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-         646±9μs          231±6μs     0.36  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-       655±200μs         227±20μs     0.35  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-       670±200μs         230±50μs     0.34  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-         665±6μs         228±10μs     0.34  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        695±30μs         238±10μs     0.34  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-         664±4μs          223±8μs     0.34  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        677±80μs         221±20μs     0.33  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'F', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
       before           after         ratio
     [3d1cb0a2]       [0386185c]
+        918±70μs       1.20±0.2ms     1.31  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'F', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
+         887±9μs      1.07±0.06ms     1.20  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'C', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
+        899±60μs      1.08±0.05ms     1.20  reduce.TimeAnyNan2D.time_anynan('float32', (1000, 1000), 'C', 0, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
+        890±20μs      1.07±0.04ms     1.20  reduce.TimeAllNan2D.time_allnan('float32', (1000, 1000), 'F', 1, 'slow') [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
```